Extract shared chart mark helpers
