+++ crates/pine-charts/src/cartesian_chart/mod.rs
@@ -13,6 +13,9 @@ use crate::legend::series_label_or_default;
 use crate::line::{
     ChartLineSeries, ChartPoint, LineChartGeometry, LineChartOptions, LineChartSample,
 };
+use crate::marks::{
+    color_or, color_or_current, key_or_index, label_or_key, rotation_transform, text_anchor,
+};
 use crate::path::{area_path, line_path};
 use crate::scale::{BandScale, LinearScale};
 use crate::svg::{format_tick, SvgAxisLabel, SvgLine, SvgTickLabel};
@@ -2545,19 +2548,6 @@ fn component_key(prefix: &str, authored_key: &str) -> String {
         .unwrap_or_else(|| prefix.into())
 }
 
-fn color_or_current(value: &str) -> String {
-    color_or(value, "currentColor")
-}
-
-fn color_or(value: &str, fallback: &str) -> String {
-    let trimmed = value.trim();
-    if trimmed.is_empty() {
-        fallback.into()
-    } else {
-        trimmed.into()
-    }
-}
-
 fn positive_or_default(value: f64, default: f64) -> f64 {
     if value.is_finite() && value > 0.0 {
         value
@@ -2582,38 +2572,6 @@ fn reference_layer(value: &str, field: &'static str) -> ChartResult<&'static str
             field,
             value: value.into(),
         }),
-    }
-}
-
-fn text_anchor(value: &str) -> &'static str {
-    match value.trim() {
-        "start" => "start",
-        "end" => "end",
-        _ => "middle",
-    }
-}
-
-fn rotation_transform(angle: f64, x: f64, y: f64) -> String {
-    if angle.abs() <= f64::EPSILON {
-        String::new()
-    } else {
-        format!("rotate({angle} {x} {y})")
-    }
-}
-
-fn label_or_key(label: &str, key: &str, index: usize) -> String {
-    if label.trim().is_empty() {
-        key_or_index("item", key, index)
-    } else {
-        label.trim().into()
-    }
-}
-
-fn key_or_index(prefix: &str, key: &str, index: usize) -> String {
-    if key.trim().is_empty() {
-        format!("{prefix}-{index}")
-    } else {
-        key.trim().into()
     }
 }
