+++ crates/pine-charts/src/layered/mod.rs
@@ -4,6 +4,9 @@ use serde::{Deserialize, Serialize};
 
 use crate::error::{finite, ChartError, ChartResult};
 use crate::geometry::Point;
+use crate::marks::{
+    color_or, color_or_current, key_or_index, label_or_key, rotation_transform, text_anchor,
+};
 use crate::path::line_path;
 
 const DEFAULT_WIDTH: f64 = 900.0;
@@ -1278,19 +1281,6 @@ fn non_negative(field: &'static str, value: f64) -> ChartResult<f64> {
     Ok(value.max(0.0))
 }
 
-fn color_or_current(value: &str) -> String {
-    color_or(value, "currentColor")
-}
-
-fn color_or(value: &str, fallback: &str) -> String {
-    let trimmed = value.trim();
-    if trimmed.is_empty() {
-        fallback.into()
-    } else {
-        trimmed.into()
-    }
-}
-
 fn reference_layer(value: &str) -> ChartResult<&'static str> {
     match value.trim() {
         "reference-foreground" | "foreground" => Ok("reference-foreground"),
@@ -1302,42 +1292,10 @@ fn reference_layer(value: &str) -> ChartResult<&'static str> {
     }
 }
 
-fn text_anchor(value: &str) -> &'static str {
-    match value.trim() {
-        "start" => "start",
-        "end" => "end",
-        _ => "middle",
-    }
-}
-
 fn icon_resolution(value: &str) -> (&'static str, &'static str) {
     match value.trim() {
         "plane" => ("plane", PLANE_ICON_PATH),
         _ => ("custom", ""),
-    }
-}
-
-fn rotation_transform(angle: f64, x: f64, y: f64) -> String {
-    if angle.abs() <= f64::EPSILON {
-        String::new()
-    } else {
-        format!("rotate({angle} {x} {y})")
-    }
-}
-
-fn label_or_key(label: &str, key: &str, index: usize) -> String {
-    if label.trim().is_empty() {
-        key_or_index("item", key, index)
-    } else {
-        label.trim().into()
-    }
-}
-
-fn key_or_index(prefix: &str, key: &str, index: usize) -> String {
-    if key.trim().is_empty() {
-        format!("{prefix}-{index}")
-    } else {
-        key.trim().into()
     }
 }
